+++ dev/cli.py
@@ -88,6 +88,7 @@ async def async_main() -> int:
         pass
 
     with commands("setup") as cmd:
+        cmd.add_argument("project", type=str, nargs="?")
         cmd.add_argument("--dev", action="store_true")
         cmd.add_argument("--local", action="store_true")
 
@@ -174,7 +175,7 @@ async def async_main() -> int:
                 mode = RepoSetupMode.DEV
             else:
                 mode = RepoSetupMode.PROD
-            setup(mode)
+            setup(mode, project=args.project)
 
         case "llmcopy":
             from dev.tasks.llmcopy import llmcopy